Makes it so it never repeats searches.
This commit adds a lot of tidying up funcitonality and makes it so the goal path is saved as well as it never searches the same node twice. Fixes many errors with the algorithm and now I just need a way to abstract the code into more macro's and functions.